rename payroll file for type
